requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Django==2.2.5
djangorestframework==3.10.3
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
html5lib==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
Pillow==6.1.0
pylint==2.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
wrapt==1.11.2
gunicorn==19.9.0
django-heroku==0.3.1
